requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
coverage==4.5.1
Flask==1.0.2
Flask-Generic-Views==0.1.1
flask-restplus==0.12.1
Flask-Views==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mirakuru==1.1.0
more-itertools==4.3.0
pluggy==0.7.1
port-for==0.4
psutil==5.5.1
psycopg2==2.7.6.1
py==1.5.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
pytest==3.7.1
pytest-cov==2.6.0
pytest-postgresql==1.3.4
python-dotenv==0.10.3
pytz==2018.5
requests==2.20.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
